+++ kolibri/core/analytics/__init__.py
@@ -1,6 +1,6 @@
 try:
-    from kolibri.core.analytics.pskolibri.common import LINUX  # noqa: F401
-    from kolibri.core.analytics.pskolibri.common import WINDOWS  # noqa: F401
+    from kolibri.utils.pskolibri.common import LINUX  # noqa: F401
+    from kolibri.utils.pskolibri.common import WINDOWS  # noqa: F401
 
     SUPPORTED_OS = True
 except NotImplementedError:
